Phase 2 step 13: Migrate TabRefreshMixin (the 51-method refresh engine, single-owner _stats producer)
